add in logics for aggregate in change_in_units.py
