+++ pyproject.toml
@@ -12,6 +12,8 @@ requests = "^2.26.0"
 python-dotenv = "^0.19.1"
 rpy2 = "^3.4.5"
 openpyxl = "^3.0.9"
+geopandas = "^0.10.2"
+Rtree = "^0.9.7"
 
 [tool.poetry.dev-dependencies]
